create app named account, do account stuff there
