+++ packages/static/js/uncompiled/app.js
@@ -20,6 +20,7 @@ function is_logged() {
 
 var csrf_token = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 const site_url = document.querySelector("meta[name='site_url']").getAttribute("content")
+const login_url = document.querySelector('meta[name="login_url"]').getAttribute('content')
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 window.axios.defaults.headers.common['X-CSRFToken'] = csrf_token;
 window.logged = is_logged()
@@ -79,7 +80,8 @@ window.store = new Vuex.Store({
     logged_in: logged_in,
     loading: true,
     api_urls: {
-        "packages": `${site_url}/api/packages`
+        "packages": `${site_url}/api/packages`,
+        'login': `${login_url}`
     }
   },
   mutations: {
