+++ packages/templates/app.html
@@ -13,6 +13,7 @@
     <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/9.12.0/styles/default.min.css">
     <meta name="csrf-token" content="{{ csrf_token }}">
     <meta name="site_url" content="{% site_url %}">
+    <meta name="login_url" content="{% url 'LoginPage' %}">
 </head>
 
 <body>
